backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
h5py==3.11.0
tensorflow-cpu==2.17.1
ultralytics>=8.2.0,<9.0.0
opencv-python-headless==4.10.0.84
openai>=1.0.0
gunicorn==22.0.0
